feat: Content 시간을 DB에 저장

+++ app/src/main/java/com/kt_study/todo_alarm/db/ContentEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class ContentEntity(
     @PrimaryKey
     val id: Int = 0,
-    val categoryId: Long = 0L,
+    val categoryId: Int = 0,
     var toDo: String = "",
     var hour: Int = 0,
     var min: Int = 0,
